REF move replacement into a new file
