+++ conda_forge_tick/migrators/__init__.py
@@ -2,13 +2,7 @@
 from .arch import ArchRebuild, OSXArm
 from .broken_rebuild import RebuildBroken
 from .conda_forge_yaml_cleanup import CondaForgeYAMLCleanup
-from .core import (
-    GraphMigrator,
-    Migrator,
-    MiniMigrator,
-    Replacement,
-    make_from_lazy_json_data,
-)
+from .core import GraphMigrator, Migrator, MiniMigrator, make_from_lazy_json_data
 from .cos7 import Cos7Config
 from .cross_compile import (
     Build2HostMigrator,
@@ -35,5 +29,6 @@ from .numpy2 import Numpy2Migrator
 from .pip_check import PipCheckMigrator
 from .pip_wheel_dep import PipWheelMigrator
 from .qt_to_qt_main import QtQtMainMigrator
+from .replacement import Replacement
 from .use_pip import PipMigrator
 from .version import Version
