+++ conda_forge_tick/migrators/core.py
@@ -1,9 +1,8 @@
-"""Classes for migrating repos"""
+"""Base classes for migrating repos"""
 
 import copy
 import datetime
 import logging
-import os
 import re
 import typing
 from typing import Any, List, Optional, Sequence, Set
@@ -226,9 +225,6 @@ class Migrator:
         graph: nx.DiGraph = None,
         effective_graph: nx.DiGraph = None,
     ):
-        if graph is not None and effective_graph is None:
-            effective_graph = _make_effective_graph(graph, self)
-
         if not hasattr(self, "_init_args"):
             self._init_args = []
 
@@ -252,6 +248,8 @@ class Migrator:
             self.graph = graph
         if effective_graph is None:
             self.effective_graph = self.graph
+        else:
+            self.effective_graph = effective_graph
 
     def to_lazy_json_data(self):
         """Serialize the migrator to LazyJson-compatible data."""
@@ -275,6 +273,12 @@ class Migrator:
         }
         data["name"] = _make_migrator_lazy_json_name(self, data)
         return data
+
+    def _reset_effective_graph(self):
+        """This method is meant to be called by an non-abstract child class at the end
+        of its __init__ method."""
+        self.effective_graph = _make_effective_graph(self.graph, self)
+        self._init_kwargs["effective_graph"] = self.effective_graph
 
     def downstream_children(
         self,
@@ -768,138 +772,6 @@ class GraphMigrator(Migrator):
             return True
 
         return False
-
-    def migrator_uid(self, attrs: "AttrsTypedDict") -> "MigrationUidTypedDict":
-        n = super().migrator_uid(attrs)
-        n["name"] = self.name
-        return n
-
-
-class Replacement(Migrator):
-    """Migrator for replacing one package with another.
-
-    Parameters
-    ----------
-    old_pkg : str
-        The package to be replaced.
-    new_pkg : str
-        The package to replace the `old_pkg`.
-    rationale : str
-        The reason the for the migration. Should be a full statement.
-    graph : nx.DiGraph, optional
-        The graph of feedstocks.
-    pr_limit : int, optional
-        The maximum number of PRs made per run of the bot.
-    check_solvable : bool, optional
-        If True, uses mamba to check if the final recipe is solvable.
-    """
-
-    migrator_version = 0
-    rerender = True
-
-    def __init__(
-        self,
-        *,
-        old_pkg: "PackageName",
-        new_pkg: "PackageName",
-        rationale: str,
-        graph: nx.DiGraph = None,
-        pr_limit: int = 0,
-        check_solvable=True,
-    ):
-        if not hasattr(self, "_init_args"):
-            self._init_args = []
-
-        if not hasattr(self, "_init_kwargs"):
-            self._init_kwargs = {
-                "old_pkg": old_pkg,
-                "new_pkg": new_pkg,
-                "rationale": rationale,
-                "graph": graph,
-                "pr_limit": pr_limit,
-                "check_solvable": check_solvable,
-            }
-
-        super().__init__(pr_limit, check_solvable=check_solvable, graph=graph)
-        self.old_pkg = old_pkg
-        self.new_pkg = new_pkg
-        self.pattern = re.compile(r"\s*-\s*(%s)(\s+|$)" % old_pkg)
-        self.packages = {old_pkg}
-        self.rationale = rationale
-        self.name = f"{old_pkg}-to-{new_pkg}"
-
-    def order(
-        self,
-        graph: nx.DiGraph,
-        total_graph: nx.DiGraph,
-    ) -> Sequence["PackageName"]:
-        """Order to run migrations in
-
-        Parameters
-        ----------
-        graph : nx.DiGraph
-            The graph of migratable PRs
-
-        Returns
-        -------
-        graph : nx.DiGraph
-            The ordered graph.
-        """
-        return graph
-
-    def filter(self, attrs: "AttrsTypedDict", not_bad_str_start: str = "") -> bool:
-        requirements = attrs.get("requirements", {})
-        rq = (
-            requirements.get("build", set())
-            | requirements.get("host", set())
-            | requirements.get("run", set())
-            | requirements.get("test", set())
-        )
-        return super().filter(attrs) or len(rq & self.packages) == 0
-
-    def migrate(
-        self, recipe_dir: str, attrs: "AttrsTypedDict", **kwargs: Any
-    ) -> "MigrationUidTypedDict":
-        with open(os.path.join(recipe_dir, "meta.yaml")) as f:
-            raw = f.read()
-        lines = raw.splitlines()
-        n = False
-        for i, line in enumerate(lines):
-            m = self.pattern.match(line)
-            if m is not None:
-                lines[i] = lines[i].replace(m.group(1), self.new_pkg)
-                n = True
-        if not n:
-            return False
-        upd = "\n".join(lines) + "\n"
-        with open(os.path.join(recipe_dir, "meta.yaml"), "w") as f:
-            f.write(upd)
-        self.set_build_number(os.path.join(recipe_dir, "meta.yaml"))
-        return super().migrate(recipe_dir, attrs)
-
-    def pr_body(self, feedstock_ctx: FeedstockContext) -> str:
-        body = super().pr_body(feedstock_ctx)
-        body = body.format(
-            "I noticed that this recipe depends on `%s` instead of \n"
-            "`%s`. %s \n"
-            "This PR makes this change."
-            "\n"
-            "Notes and instructions for merging this PR:\n"
-            "1. Make sure that the recipe can indeed only depend on `%s`. \n"
-            "2. Please merge the PR only after the tests have passed. \n"
-            "3. Feel free to push to the bot's branch to update this PR if "
-            "needed. \n" % (self.old_pkg, self.new_pkg, self.rationale, self.new_pkg),
-        )
-        return body
-
-    def commit_message(self, feedstock_ctx: FeedstockContext) -> str:
-        return f"use {self.new_pkg} instead of {self.old_pkg}"
-
-    def pr_title(self, feedstock_ctx: FeedstockContext) -> str:
-        return f"Suggestion: depend on {self.new_pkg} instead of {self.old_pkg}"
-
-    def remote_branch(self, feedstock_ctx: FeedstockContext) -> str:
-        return f"{self.old_pkg}-to-{self.new_pkg}-migration-{self.migrator_version}"
 
     def migrator_uid(self, attrs: "AttrsTypedDict") -> "MigrationUidTypedDict":
         n = super().migrator_uid(attrs)
